+++ custom_components/wiser/const.py
@@ -47,6 +47,10 @@ CONF_RESTORE_MANUAL_TEMP_OPTION = "restore_manual_temp_option"
 
 # Custom Attributes
 ATTR_TIME_PERIOD = "time_period"
+ATTR_FILENAME = "filename"
+ATTR_TO_ENTITY_ID = "to_entity_id"
+ATTR_SCHEDULE_ID = "schedule_id"
+
 
 # Signal icons
 SIGNAL_STRENGTH_ICONS = {
